Table of summary statistics and model estimates for laboratory values with continuous outcomes

labtable is used to plot a table that summarises laboratory values with continuous outcomes for baseline and each post-baseline timepoint by treatment arms using the following summary statistics:

Mean (SD)

Median (IQR)

Number of missing n

Treatment effect estimate (95% CI)

Usage

labtable(
  data,
  control,
  intervention_levels,
  lab_test = "lab_test",
  visit = "visit",
  id = "id",
  arm = "arm",
  aval = "aval",
  treatment_effect_estimate = TRUE,
  model_formula = "aval ~ arm + (1|id)",
  mean = TRUE,
  median = TRUE,
  n_missing = TRUE,
  control_name = NULL,
  intervention_names = NULL,
  mean_dp = 1,
  SD_dp = 1,
  median_dp = 1,
  IQR_dp = 1,
  estimate_sf = 3,
  CI_sf = 3,
  save_image_path = NULL,
  save_docx_path = NULL
)

Arguments

data

dataframe with lab_test, visit, id, arm and aval columns (optional: variables to be included in the model)

control

factor level of control arm

intervention_levels

vector of factor levels for interventions

lab_test

name of lab_test column

visit

name of visit column

id

name of id column

arm

name of arm column

aval

name of aval column

treatment_effect_estimate

a logical value whether to include treatment effect estimate and 95% CI column

model_formula

formula for linear mixed model to estimate treatment effect and 95% CI

mean

a logical value whether to include mean and SD row in summary table

median

a logical value whether to include median and IQR row in summary table

n_missing

a logical value whether to include number of missing n row in summary table

control_name

name of control arm

intervention_names

vector of names for interventions

mean_dp

number of decimal places for mean

SD_dp

number of decimal places for standard deviation

median_dp

number of decimal places for median

IQR_dp

number of decimal places for IQR

estimate_sf

number of significant figures for treatment effect estimate

CI_sf

number of significant figures for 95% CI

save_image_path

file path to save table as image

save_docx_path

file path to save table as docx

Value

flextable of laboratory values summary for baseline and each post-baseline timepoint by treatment arms

Treatment Effect Estimate

The treatment effect and its 95% confidence interval is estimated via a linear mixed model with laboratory measurement as response. The formula of the model can be specified through the model_formula argument. The default model is aval ~ arm + (1|id).

Examples

labtable(lab2, control="Placebo", intervention_levels=c("Intervention"))

Intervention (N₁=66)

Placebo (N₂=69)

Treatment effect estimate (95% CI)

Alanine Aminotransferase (IU/L)

Week 0

Mean (SD)

56.6 (32)

61.1 (34.9)

-1.39 (-6.33, 3.53)

Median (IQR)

55 (52.6)

56.2 (58)

Number of missing n

1

0

Week 4

Mean (SD)

61.6 (35.8)

61.6 (35.1)

Median (IQR)

56.6 (69.5)

61.8 (59.1)

Number of missing n

3

2

Week 8

Mean (SD)

57.6 (32.5)

64.4 (32.3)

Median (IQR)

53.3 (50.5)

68.8 (58.1)

Number of missing n

3

1

Week 12

Mean (SD)

60.8 (31.2)

61.7 (34.3)

Median (IQR)

64.9 (47.4)

59.3 (55.2)

Number of missing n

3

4

Week 16

Mean (SD)

NA (NA)

92.5 (NA)

Median (IQR)

NA (NA)

92.5 (0)

Number of missing n

66

68

Week 20

Mean (SD)

61.3 (35)

55.4 (30.8)

Median (IQR)

68.4 (63.3)

60.6 (50.6)

Number of missing n

3

4

Week 24

Mean (SD)

60.1 (31.3)

61.2 (32.6)

Median (IQR)

60.9 (53.9)

60.1 (52.9)

Number of missing n

4

5

Lymphocytes (GI/L)

Week 0

Mean (SD)

3.6 (1.9)

3.2 (2)

-0.156 (-0.423, 0.111)

Median (IQR)

3.5 (3.3)

3 (3.2)

Number of missing n

1

2

Week 4

Mean (SD)

3.6 (1.8)

3.8 (2.1)

Median (IQR)

3.6 (2.5)

3.7 (3.8)

Number of missing n

4

4

Week 8

Mean (SD)

3.8 (2.1)

4.2 (1.9)

Median (IQR)

3.5 (3.9)

4.4 (3.6)

Number of missing n

3

2

Week 12

Mean (SD)

3.8 (1.9)

4.1 (1.9)

Median (IQR)

4 (2.9)

4.3 (3.1)

Number of missing n

3

3

Week 16

Mean (SD)

3.9 (2.1)

4.1 (2)

Median (IQR)

3.8 (3.9)

4.2 (3.4)

Number of missing n

2

5

Week 20

Mean (SD)

3.9 (1.8)

4 (2)

Median (IQR)

4 (2.5)

4.1 (3.3)

Number of missing n

4

3

Week 24

Mean (SD)

3.5 (1.9)

3.9 (2.2)

Median (IQR)

3.4 (3.4)

3.7 (3.9)

Number of missing n

4

4

Neutrophils (GI/L)

Week 0

Mean (SD)

12.6 (6.5)

12 (7)

0.344 (-0.589, 1.28)

Median (IQR)

13.8 (11.5)

11.4 (12.5)

Number of missing n

1

2

Week 4

Mean (SD)

12.9 (6.4)

11.3 (6.8)

Median (IQR)

12.9 (11.5)

11 (12.3)

Number of missing n

4

4

Week 8

Mean (SD)

11.8 (7.1)

11.6 (6.8)

Median (IQR)

11.8 (13.2)

10.4 (13)

Number of missing n

3

2

Week 12

Mean (SD)

12.1 (6.8)

12.2 (6.7)

Median (IQR)

11.1 (12.5)

11.1 (12.5)

Number of missing n

3

3

Week 16

Mean (SD)

12.7 (6.8)

11.5 (7.2)

Median (IQR)

13 (11.8)

12.2 (13.1)

Number of missing n

2

5

Week 20

Mean (SD)

11.2 (6.7)

12.1 (6.7)

Median (IQR)

10.3 (11.4)

12.2 (10.3)

Number of missing n

4

3

Week 24

Mean (SD)

12.4 (6.8)

12.6 (6.2)

Median (IQR)

13.1 (12)

13.3 (11.5)

Number of missing n

4

4